requirements.txt
streamlit
numpy
Pillow
opencv-python-headless
tensorflow
scikit-learn
joblib